refactor: move components\Search\Filter to components\Filter
