+++ src/pages/NervosDao/index.tsx
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query'
 import Content from '../../components/Content'
 import { DaoContentPanel, DaoTabBarPanel } from './styled'
 import DaoTransactions from './DaoTransactions'
-import Filter from '../../components/Search/Filter'
+import Filter from '../../components/Filter'
 import DepositorRank from './DepositorRank'
 import { usePaginationParamsInPage, useSearchParams } from '../../hooks'
 import DaoOverview from './DaoOverview'
