+++ src/pages/NftCollectionInfo/index.tsx
@@ -6,7 +6,7 @@ import Content from '../../components/Content'
 import NftCollectionOverview from './NftCollectionOverview'
 import NftCollectionTransfers from './NftCollectionTransfers'
 import NftCollectionInventory from './NftCollectionInventory'
-import Filter from '../../components/Search/Filter'
+import Filter from '../../components/Filter'
 import { ReactComponent as FilterIcon } from '../../assets/filter_icon.svg'
 import { ReactComponent as SelectedCheckIcon } from '../../assets/selected_check_icon.svg'
 import { explorerService } from '../../services/ExplorerService'
