+++ src/pages/UDT/index.tsx
@@ -6,7 +6,7 @@ import Content from '../../components/Content'
 import { UDTContentPanel, UDTTransactionTitlePanel } from './styled'
 import UDTComp, { UDTOverviewCard } from './UDTComp'
 import { usePaginationParamsInPage, useSearchParams, useUpdateSearchParams } from '../../hooks'
-import Filter from '../../components/Search/Filter'
+import Filter from '../../components/Filter'
 import { localeNumberString } from '../../utils/number'
 import { explorerService } from '../../services/ExplorerService'
 import { deprecatedAddrToNewAddr } from '../../utils/util'
